+++ src/Ext/SWType/Body.h
@@ -2,6 +2,8 @@
 #include <Ext/Building/Body.h>
 #include <Utilities/Container.h>
 #include <Utilities/TemplateDef.h>
+
+#include <New/Type/SWSignalTypeClass.h>
 #include <New/Type/Affiliated/TypeConvertGroup.h>
 
 class SWTypeExt
@@ -30,8 +32,10 @@ public:
 		Valueable<bool> SW_Unstoppable;
 		ValueableVector<TechnoTypeClass*> SW_Inhibitors;
 		Valueable<bool> SW_AnyInhibitor;
+		Valueable<AffectedHouse> SW_InhibitorsHouse;
 		ValueableVector<TechnoTypeClass*> SW_Designators;
 		Valueable<bool> SW_AnyDesignator;
+		Valueable<AffectedHouse> SW_DesignatorsHouse;
 		Valueable<double> SW_RangeMinimum;
 		Valueable<double> SW_RangeMaximum;
 		Valueable<int> SW_Shots;
@@ -43,6 +47,9 @@ public:
 		Valueable<bool> SW_InitialReady;
 		ValueableIdx<SuperWeaponTypeClass> SW_PostDependent;
 		Valueable<int> SW_MaxCount;
+
+		ValueableVector<SWSignalTypeClass*> SW_DesignateTypes;
+		ValueableVector<SWSignalTypeClass*> SW_InhibitTypes;
 
 		Valueable<CSFText> Message_CannotFire;
 		Valueable<CSFText> Message_InsufficientFunds;
@@ -119,8 +126,10 @@ public:
 			, SW_Unstoppable { false }
 			, SW_Inhibitors {}
 			, SW_AnyInhibitor { false }
+			, SW_InhibitorsHouse { AffectedHouse::Enemies }
 			, SW_Designators { }
 			, SW_AnyDesignator { false }
+			, SW_DesignatorsHouse { AffectedHouse::Owner }
 			, SW_RangeMinimum { -1.0 }
 			, SW_RangeMaximum { -1.0 }
 			, SW_RequiredHouses { 0xFFFFFFFFu }
@@ -130,6 +139,8 @@ public:
 			, SW_InitialReady { false }
 			, SW_PostDependent {}
 			, SW_MaxCount { -1 }
+			, SW_DesignateTypes {}
+			, SW_InhibitTypes {}
 			, SW_Shots { -1 }
 			, Message_CannotFire {}
 			, Message_InsufficientFunds {}
@@ -184,10 +195,10 @@ public:
 		{ }
 
 		// Ares 0.A functions
-		bool IsInhibitor(HouseClass* pOwner, TechnoClass* pTechno) const;
+		//bool IsInhibitor(HouseClass* pOwner, TechnoClass* pTechno) const;
 		bool HasInhibitor(HouseClass* pOwner, const CellStruct& coords) const;
 		bool IsInhibitorEligible(HouseClass* pOwner, const CellStruct& coords, TechnoClass* pTechno) const;
-		bool IsDesignator(HouseClass* pOwner, TechnoClass* pTechno) const;
+		//bool IsDesignator(HouseClass* pOwner, TechnoClass* pTechno) const;
 		bool HasDesignator(HouseClass* pOwner, const CellStruct& coords) const;
 		bool IsDesignatorEligible(HouseClass* pOwner, const CellStruct& coords, TechnoClass* pTechno) const;
 		bool IsLaunchSiteEligible(const CellStruct& Coords, BuildingClass* pBuilding, bool ignoreRange) const;