+++ src/Ext/TechnoType/Body.h
@@ -7,6 +7,7 @@
 #include <New/Type/LaserTrailTypeClass.h>
 #include <New/Type/DigitalDisplayTypeClass.h>
 #include <New/Type/SelectBoxTypeClass.h>
+#include <New/Type/SWSignalTypeClass.h>
 #include <New/Type/Affiliated/InterceptorTypeClass.h>
 #include <New/Type/Affiliated/PassengerDeletionTypeClass.h>
 #include <New/Type/Affiliated/DroppodTypeClass.h>
@@ -345,6 +346,9 @@ public:
 		Valueable<Leptons> KeepTargetOnMove_ExtraDistance;
 
 		Valueable<int> Power;
+
+		ValueableVector<SWSignalTypeClass*> DesignateTypes;
+		ValueableVector<SWSignalTypeClass*> InhibitTypes;
 
 		Nullable<bool> AllowAirstrike;
 
@@ -783,6 +787,9 @@ public:
 			, KeepTargetOnMove_ExtraDistance { Leptons(0) }
 
 			, Power { }
+
+			, DesignateTypes { }
+			, InhibitTypes { }
 
 			, AllowAirstrike { }