+++ src/Misc/Hooks.Ares.cpp
@@ -3,6 +3,9 @@
 
 #include <Ext/Building/Body.h>
 #include <Ext/Sidebar/Body.h>
+#include <Ext/Techno/Body.h>
+#include <Ext/SWType/Body.h>
+#include <Ext/SWType/Ares/NewSWType.h>
 #include <Ext/EBolt/Body.h>
 
 #include <New/Entity/Ares/RadarJammerClass.h>
@@ -37,7 +40,6 @@ static bool __stdcall ConvertToType(TechnoClass* pThis, TechnoTypeClass* pToType
 static TechnoTypeClass* __fastcall ShowPromoteAnim(TechnoClass* pThis)
 {
 	TechnoExt::ShowPromoteAnim(pThis);
-
 	return pThis->GetTechnoType();
 }
 
@@ -45,6 +47,8 @@ static WeaponStruct* __fastcall GetLaserWeapon(BuildingClass* pThis)
 {
 	return BuildingExt::GetLaserWeapon(pThis);
 }
+
+_GET_FUNCTION_ADDRESS(AresNewSWType::GetTargetingData, AresNewSWType_GetTargetingData_GetAddr)
 
 static EBolt* __stdcall CreateEBolt(WeaponTypeClass** pWeaponData)
 {
@@ -146,6 +150,9 @@ void Apply_Ares3_0_Patches()
   
 	// Redirect Ares's function to our implementation:
 	Patch::Apply_LJMP(AresHelper::AresBaseAddress + 0x112D0, &BuildingExt::KickOutClone);
+  
+  // Redirect Ares' NewSWType::GetTargetingData() to our implementation:
+	Patch::Apply_LJMP(AresHelper::AresBaseAddress + 0x6D1E0, AresNewSWType_GetTargetingData_GetAddr());
 
 	// Redirect Ares's TechnoTypeExt::ExtData::CameoIsElite() to our implementation:
 	Patch::Apply_LJMP(AresHelper::AresBaseAddress + 0x3D800, &CameoIsVeteran);
@@ -221,6 +228,9 @@ void Apply_Ares3_0p1_Patches()
   
 	// Redirect Ares's function to our implementation:
 	Patch::Apply_LJMP(AresHelper::AresBaseAddress + 0x11860, &BuildingExt::KickOutClone);
+  
+  // Redirect Ares' NewSWType::GetTargetingData() to our implementation:
+	Patch::Apply_LJMP(AresHelper::AresBaseAddress + 0x6E1F0, AresNewSWType_GetTargetingData_GetAddr());
 
 	// Redirect Ares's TechnoTypeExt::ExtData::CameoIsElite() to our implementation:
 	Patch::Apply_LJMP(AresHelper::AresBaseAddress + 0x3E210, &CameoIsVeteran);