+++ src/New/Type/SWSignalTypeClass.h
@@ -1,0 +1,24 @@
+#pragma once
+
+#include <Utilities/Enumerable.h>
+#include <Utilities/TemplateDef.h>
+
+class SWSignalTypeClass final : public Enumerable<SWSignalTypeClass>
+{
+public:
+	Nullable<int> Range {};
+	Nullable<AffectedHouse> Affects {};
+	Valueable<bool> Powered { false };
+	Valueable<bool> StopInTemporal { false };
+
+	SWSignalTypeClass(const char* pTitle = NONE_STR) : Enumerable<SWSignalTypeClass>(pTitle)
+	{ }
+
+	void LoadFromINI(CCINIClass* pINI);
+	void LoadFromStream(PhobosStreamReader& stm);
+	void SaveToStream(PhobosStreamWriter& stm);
+
+private:
+	template <typename T>
+	void Serialize(T& stm);
+};